+++ noodles-bam/src/writer/record.rs
@@ -234,7 +234,7 @@ pub(super) fn write_sequence<W>(writer: &mut W, sequence: &sam::record::Sequence
 where
     W: Write,
 {
-    for chunk in sequence.chunks(2) {
+    for chunk in sequence.as_ref().chunks(2) {
         let l = chunk[0];
         let r = chunk.get(1).copied().unwrap_or(Base::Eq);
         let b = encode_base(l) << 4 | encode_base(r);